+++ lib/mongoid/components.rb
@@ -8,14 +8,11 @@ module Mongoid
     included do
       extend ActiveModel::Translation
       extend Mongoid::Finders
-
-      class_attribute :paranoid
     end
 
     include ActiveModel::Conversion
-    include ActiveModel::MassAssignmentSecurity
+    include ActiveModel::ForbiddenAttributesProtection
     include ActiveModel::Naming
-    include ActiveModel::Observing
     include ActiveModel::Serializers::JSON
     include ActiveModel::Serializers::Xml
     include Mongoid::Atomic