+++ lib/mongoid/fields/standard.rb
@@ -168,18 +168,6 @@ module Mongoid
         @type ||= options[:type] || Object
       end
 
-      # Is this field included in versioned attributes?
-      #
-      # @example Is the field versioned?
-      #   field.versioned?
-      #
-      # @return [ true, false ] If the field is included in versioning.
-      #
-      # @since 2.1.0
-      def versioned?
-        @versioned ||= (options[:versioned].nil? ? true : options[:versioned])
-      end
-
       private
 
       # Get the name of the default method for this field.