+++ lib/mongoid/validations/uniqueness.rb
@@ -170,7 +170,6 @@ module Mongoid
           name = document.database_field_name(item)
           criteria = criteria.where(item => document.attributes[name])
         end
-        criteria = criteria.where(deleted_at: nil) if document.paranoid?
         criteria
       end