+++ CHANGELOG.md
@@ -3,6 +3,26 @@
 For instructions on upgrading to newer versions, visit
 [mongoid.org](http://mongoid.org/en/mongoid/docs/upgrading.html).
 
+## 4.0.0 (branch: 4.0.0-dev)
+
+### Major Changes (Backwards Incompatible)
+
+* Mongoid now requires Active Model 4 or higher.
+
+* \#2659 `Mongoid::Railtie` now properly uses only one initializer and
+  the name has changed to `mongoid.load-config`.
+
+* \#2656 `rake db:reseed` is now `rake db:reset` (Arthur Neves)
+
+* \#2648 `Boolean` becomes `Mongoid::Boolean` to avoid polluting the global
+  namespace with a commonly used class by other libraries.
+
+* \#2433 `Mongoid::Paranoia` has been removed.
+
+* \#2432 `Mongoid::Versioning` has been removed.
+
+* \#2200 Mass assignment security now mirrors Rails 4's behavior.
+
 ## 3.1.0 (branch: master)
 
 ### New Features